Add 'adjusted heuristic' deduction rule
The 'linear interpolation' retirement deduction rule suffers from the problem that many initial guesses are likely to not lead to a solution, which the optimizing solver is not equipped to handle.

The new rule is the earlier 'heuristic' rule, which will normally find a solution if one exists, but with an additional adjustable offset for the optimizer to work with.